implement versioning: v1 as default  - add v2 for testing
