+++ app/controllers/v2/todos_controller.rb
@@ -1,0 +1,7 @@
+module V2
+  class TodosController < ApplicationController
+    def index
+      json_response({ message: "Version 2 implementation" })
+    end
+  end
+end
